Creando el Home (3/25)

+++ src/app/components/products/products.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 
 import { CreateProductDTO, Product, UpdateProductDTO } from '../../models/product.model';
 
@@ -10,23 +10,14 @@ import { ProductsService } from '../../services/products.service';
   templateUrl: './products.component.html',
   styleUrls: ['./products.component.scss']
 })
-export class ProductsComponent implements OnInit {
+export class ProductsComponent {
 
+  @Input() products: Product[] = [];
+  @Output() loadMore: EventEmitter<string> = new EventEmitter();
   myShoppingCart: Product[] = [];
   total = 0;
-  products: Product[] = [];
   showProductDetail: boolean = false;
-  productChosen: Product = {
-    id: '',
-    price: 0,
-    images: [],
-    title: '',
-    category: {
-      id: '',
-      name: '',
-    },
-    description: ''
-  };
+  productChosen: Product | null = null;
   limit = 10;
   offset = 0;
   statusDetail: 'loading' | 'success' | 'error' | 'init' = 'init';
@@ -37,19 +28,6 @@ export class ProductsComponent implements OnInit {
     private productsService: ProductsService
   ) {
     this.myShoppingCart = this.storeService.getShoppingCart();
-  }
-
-  ngOnInit(): void {
-    this.productsService.getAllProducts(10, 0)
-      .subscribe(data => {
-        this.products = data;
-        this.offset += this.limit;
-      });
-    // this.productsService.getAllProducts(10, 0)
-    // .subscribe(data => {
-    //   this.products = data;
-    //   this.offset += this.limit;
-    // });
   }
 
   onAddToShoppingCart(product: Product) {
@@ -88,31 +66,32 @@ export class ProductsComponent implements OnInit {
   }
 
   updateProduct() {
-    const changes: UpdateProductDTO = {
-      title: 'change title',
+    if (this.productChosen) {
+      const changes: UpdateProductDTO = {
+        title: 'change title',
+      }
+      const id = this.productChosen.id;
+      this.productsService.update(id, changes)
+        .subscribe(data => {
+          const productIndex = this.products.findIndex(item => item.id === this.productChosen?.id);
+          this.products[productIndex] = data;
+          this.productChosen = data;
+        });
     }
-    const id = this.productChosen.id;
-    this.productsService.update(id, changes)
-      .subscribe(data => {
-        const productIndex = this.products.findIndex(item => item.id === this.productChosen.id);
-        this.products[productIndex] = data;
-        this.productChosen = data;
-      });
   }
 
   deleteProduct() {
-    const id = this.productChosen.id;
-    this.productsService.delete(id).subscribe(() => {
-      const productIndex = this.products.findIndex(item => item.id === this.productChosen.id);
-      this.products.splice(productIndex, 1);
-      this.showProductDetail = false;
-    })
+    if (this.productChosen) {
+      const id = this.productChosen.id;
+      this.productsService.delete(id).subscribe(() => {
+        const productIndex = this.products.findIndex(item => item.id === this.productChosen?.id);
+        this.products.splice(productIndex, 1);
+        this.showProductDetail = false;
+      })
+    }
   }
 
-  loadMore() {
-    this.productsService.getProductsByPage(this.limit, this.offset).subscribe(data => {
-      this.products = this.products.concat(data);
-      this.offset += this.limit;
-    })
+  onLoadMore() {
+    this.loadMore.emit();
   }
 }
